Reorganise homepage sections into grouped   subdirectories
  Group flat numbered files into hero/,
  features/, about/, footer/
  subfolders. No content changes — imports
  and relative paths updated.    Merge develop
