+++ app/(home)/page.tsx
@@ -2,16 +2,16 @@
 
 import { useState } from "react";
 import { usePromptStore } from "@/app/store/promptStore";
-import LandingHero from "./sections/1_Hero";
-import PromptMarquee from "./sections/2_PromptMarquee";
-import FeaturesTabsSection from "./sections/4_FeaturesTabs";
-import SupportWorkTabsSection from "./sections/5_SupportWorkTabs";
-import HowItWorksSection from "./sections/6_HowItWorks";
-import LatestUpdatesSection from "./sections/7_LatestUpdates";
-import FutureOfMonitoringSection from "./sections/8_FutureOfMonitoring";
-import TeamSection from "./sections/9_TeamSection";
-import CTASection from "./sections/11_CTA";
-import FooterSection from "./sections/12_Footer";
+import LandingHero from "./sections/hero/Hero";
+import PromptMarquee from "./sections/hero/PromptMarquee";
+import FeaturesTabsSection from "./sections/features/FeaturesTabs";
+import SupportWorkTabsSection from "./sections/features/SupportWorkTabs";
+import HowItWorksSection from "./sections/features/HowItWorks";
+import LatestUpdatesSection from "./sections/footer/LatestUpdates";
+import FutureOfMonitoringSection from "./sections/about/FutureOfMonitoring";
+import TeamSection from "./sections/about/TeamSection";
+import CTASection from "./sections/footer/CTA";
+import FooterSection from "./sections/footer/Footer";
 
 export default function LandingPage() {
   const [promptIndex, setPromptIndex] = useState(0);
