+++ app/(home)/sections/about/FutureOfMonitoring.tsx
