+++ app/(home)/sections/about/NewEraQuote.tsx
