+++ app/(home)/sections/about/TeamSection.tsx
