+++ app/(home)/sections/features/HowItWorks.tsx
